+++ extract/singer/marketo/tap_marketo/client.py
@@ -1,0 +1,9 @@
+import aiohttp
+
+
+class MarketoClient(object):
+    
+
+
+if __name__ == '__main__':
+    Fire(marketo_client)
